requirements.txt
nose
numpy
librosa
scikit-maad
scipy
sounddevice
typing_extensions